pyproject.toml
[project]
name = "mcp-google-agent-platform-docs"
version = "0.1.0"
description = "MCP server providing Google AI platform documentation (Gemini Enterprise Agent Platform + Vertex AI) to AI agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["mcp", "google-ai", "gemini", "vertex-ai", "documentation", "agents", "geap"]
authors = [
    {name = "rrromochka"},
]

dependencies = [
    "mcp[cli]>=1.20.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.13.0",
    "pyyaml>=6.0",
]

[project.scripts]
mcp-google-agent-platform-docs = "mcp_google_agent_platform_docs.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_google_agent_platform_docs"]
